[RF][Structure][_Languages][POM refactoring]<sabinasaad>
