+++ src/test/java/pages/browse_languages/languages/LanguageAlgol60Page.java
@@ -4,19 +4,14 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-public class LanguageAlgol60Page {
+public class LanguageAlgol60Page extends _LanguagesPage{
     private WebDriver driver;
 
     private final By LINK_ALGOL_60_LANGUAGE_SITE = By.xpath("//a[@href='language-algol60-1367.html']");
 
-    public LanguageAlgol60Page(WebDriver existingDriver) {
+    public LanguageAlgol60Page(WebDriver driver) {
 
-        this.driver = existingDriver;
-    }
-
-    protected WebDriver getDriver() {
-
-        return driver;
+        super(driver);
     }
 
     public WebElement getLanguageAlgol60Website() {
