+++ src/test/java/pages/browse_languages/languages/LanguageMathematicaPage.java
@@ -4,20 +4,12 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-public class LanguageMathematicaPage {
-
-    private WebDriver driver;
+public class LanguageMathematicaPage extends _LanguagesPage {
 
     private final By TD_TABLE_BRENTON_BOSTICK = By.xpath("//table[@style='margin: 7px; padding: 0;;']/tbody[1]/tr[2]/td[2]");
 
-    public LanguageMathematicaPage(WebDriver existingDriver) {
-
-        this.driver = existingDriver;
-    }
-
-    protected WebDriver getDriver() {
-
-        return driver;
+    public LanguageMathematicaPage(WebDriver driver) {
+        super(driver);
     }
 
     public WebElement getTdTableBrentonBostick() {
